Move float sampling from Rand to distributions
Move default sampling to Default distribution
Make Open01 and Closed01 distributions instead of wrappers
