+++ src/rand_impls.rs
@@ -109,48 +109,6 @@ impl Rand for u128 {
 }
 
 
-macro_rules! float_impls {
-    ($mod_name:ident, $ty:ty, $mantissa_bits:expr, $method_name:ident) => {
-        mod $mod_name {
-            use {Rand, Rng, Open01, Closed01};
-
-            // 1.0 / epsilon
-            const SCALE: $ty = (1u64 << $mantissa_bits) as $ty;
-
-            impl Rand for $ty {
-                /// Generate a floating point number in the half-open
-                /// interval `[0,1)`.
-                ///
-                /// See `Closed01` for the closed interval `[0,1]`,
-                /// and `Open01` for the open interval `(0,1)`.
-                #[inline]
-                fn rand<R: Rng>(rng: &mut R) -> $ty {
-                    rng.$method_name()
-                }
-            }
-            impl Rand for Open01<$ty> {
-                #[inline]
-                fn rand<R: Rng>(rng: &mut R) -> Open01<$ty> {
-                    // add 0.5 * epsilon, so that smallest number is
-                    // greater than 0, and largest number is still
-                    // less than 1, specifically 1 - 0.5 * epsilon.
-                    Open01(rng.$method_name() + 0.5 / SCALE)
-                }
-            }
-            impl Rand for Closed01<$ty> {
-                #[inline]
-                fn rand<R: Rng>(rng: &mut R) -> Closed01<$ty> {
-                    // rescale so that 1.0 - epsilon becomes 1.0
-                    // precisely.
-                    Closed01(rng.$method_name() * SCALE / (SCALE - 1.0))
-                }
-            }
-        }
-    }
-}
-float_impls! { f64_rand_impls, f64, 52, next_f64 }
-float_impls! { f32_rand_impls, f32, 23, next_f32 }
-
 impl Rand for char {
     #[inline]
     fn rand<R: Rng>(rng: &mut R) -> char {
@@ -242,116 +200,6 @@ impl<T:Rand> Rand for Option<T> {
             Some(rng.gen())
         } else {
             None
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use impls;
-    use {RngCore, Rng, Open01, Closed01};
-    
-    const EPSILON32: f32 = ::core::f32::EPSILON;
-    const EPSILON64: f64 = ::core::f64::EPSILON;
-
-    struct ConstantRng(u64);
-    impl RngCore for ConstantRng {
-        fn next_u32(&mut self) -> u32 {
-            let ConstantRng(v) = *self;
-            v as u32
-        }
-        fn next_u64(&mut self) -> u64 {
-            let ConstantRng(v) = *self;
-            v
-        }
-
-        fn fill_bytes(&mut self, dest: &mut [u8]) {
-            impls::fill_bytes_via_u64(self, dest)
-        }
-    }
-
-    #[test]
-    fn floating_point_edge_cases() {
-        let mut zeros = ConstantRng(0);
-        assert_eq!(zeros.gen::<f32>(), 0.0);
-        assert_eq!(zeros.gen::<f64>(), 0.0);
-        
-        let mut one = ConstantRng(1);
-        assert_eq!(one.gen::<f32>(), EPSILON32);
-        assert_eq!(one.gen::<f64>(), EPSILON64);
-        
-        let mut max = ConstantRng(!0);
-        assert_eq!(max.gen::<f32>(), 1.0 - EPSILON32);
-        assert_eq!(max.gen::<f64>(), 1.0 - EPSILON64);
-    }
-
-    #[test]
-    fn fp_closed_edge_cases() {
-        let mut zeros = ConstantRng(0);
-        let Closed01(zero32) = zeros.gen::<Closed01<f32>>();
-        let Closed01(zero64) = zeros.gen::<Closed01<f64>>();
-        assert_eq!(zero32, 0.0);
-        assert_eq!(zero64, 0.0);
-        
-        let mut one = ConstantRng(1);
-        let Closed01(one32) = one.gen::<Closed01<f32>>();
-        let Closed01(one64) = one.gen::<Closed01<f64>>();
-        assert!(EPSILON32 < one32 && one32 < EPSILON32 * 1.01);
-        assert!(EPSILON64 < one64 && one64 < EPSILON64 * 1.01);
-        
-        let mut max = ConstantRng(!0);
-        let Closed01(max32) = max.gen::<Closed01<f32>>();
-        let Closed01(max64) = max.gen::<Closed01<f64>>();
-        assert_eq!(max32, 1.0);
-        assert_eq!(max64, 1.0);
-    }
-
-    #[test]
-    fn fp_open_edge_cases() {
-        let mut zeros = ConstantRng(0);
-        let Open01(zero32) = zeros.gen::<Open01<f32>>();
-        let Open01(zero64) = zeros.gen::<Open01<f64>>();
-        assert_eq!(zero32, 0.0 + EPSILON32 / 2.0);
-        assert_eq!(zero64, 0.0 + EPSILON64 / 2.0);
-        
-        let mut one = ConstantRng(1);
-        let Open01(one32) = one.gen::<Open01<f32>>();
-        let Open01(one64) = one.gen::<Open01<f64>>();
-        assert!(EPSILON32 < one32 && one32 < EPSILON32 * 2.0);
-        assert!(EPSILON64 < one64 && one64 < EPSILON64 * 2.0);
-        
-        let mut max = ConstantRng(!0);
-        let Open01(max32) = max.gen::<Open01<f32>>();
-        let Open01(max64) = max.gen::<Open01<f64>>();
-        assert_eq!(max32, 1.0 - EPSILON32 / 2.0);
-        assert_eq!(max64, 1.0 - EPSILON64 / 2.0);
-    }
-
-    #[test]
-    fn rand_open() {
-        // this is unlikely to catch an incorrect implementation that
-        // generates exactly 0 or 1, but it keeps it sane.
-        let mut rng = ::test::rng(501);
-        for _ in 0..1_000 {
-            // strict inequalities
-            let Open01(f) = rng.gen::<Open01<f64>>();
-            assert!(0.0 < f && f < 1.0);
-
-            let Open01(f) = rng.gen::<Open01<f32>>();
-            assert!(0.0 < f && f < 1.0);
-        }
-    }
-
-    #[test]
-    fn rand_closed() {
-        let mut rng = ::test::rng(502);
-        for _ in 0..1_000 {
-            // strict inequalities
-            let Closed01(f) = rng.gen::<Closed01<f64>>();
-            assert!(0.0 <= f && f <= 1.0);
-
-            let Closed01(f) = rng.gen::<Closed01<f32>>();
-            assert!(0.0 <= f && f <= 1.0);
         }
     }
 }
